Bug #13964
Fix the bug which is in the list of the connected users. In this page,
the SessionManagement service is invoked to get all the opened user
sessions and for each of them the related user. But, the user in the
session is the one that has been fetched from the database during the
login process, before opening really the session. So, the last login
date attribute of the user is the one of his previous latest session and
not the current one that is set once the session is opened. The fix was
to reuse the user in the newly opened session when setting its login
date attribute before saving his data in the database.