Feature #14962
In the Silverpeas backoffice, the two types of user groups are now distinguished.
Actions on application groups are now forbidden (and hidden) because
those groups are managed by the applications that create them.
When rendering the groups (either in the backoffice or in the user
panel), the groups managed by a user domain are rendered first, then the
groups managed by applications.