Feature #14962
Add a mechanism for admin to create explicitly a community space by
keeping the Silverpeas component that motorizes a space as a
community space.
The community space is created like any collaborative space but with
additional operations: the space is marked as a community one and for
doing it has some of its properties set with a predefined value (like
the access rights inheritance blocked). Then the creation of the
component instance dedicated to manage the community membership is
invoked.

To allow this, a new property, isCommunitySpace, is added to the
SpaceInst class and a new interface, CommunityFactory, is defined. This
interface has to be implemented by the already existing Silverpeas
commponent Community (defined in the Silverpeas Components project) so
that Silverpeas Core is decoupled to any code that motorize the
community feature. This feature is enabled if and only if there is an
implementation of the CommunityFactory interface.