Bug #14969
For the subscriptions on nodes (folders in Kmelia for example), take
into account the move of nodes. In the case the node is move into
another component instance, then any subscriptions on that node are
renewed on the node.

In order to do this, add the notification on nodes move.