Bug #14975
In a form, a publication picker field is rendered as text in read-only mode.
the TextDisplayer used for doing encodes the value for HTML for any
fields other than file and date field. Hence the bug with publication
URL fields. So the fix is avoid the HTML encoding for such type of fields.
(Publication URLs field are the read-only rendering of publications
picker fiel.d)