Bug #15077
When files are uploaded in a Kmelia instance, the parameters (like the
kmelia instance id, the publication name, and so on) are passed directly
in the URL (hu ???). This is why publication name, description and
keywords are URL encoded. So, in the servlet targeted by the URL, these
parameters has to be decoded before being used to build the publication.